pyproject.toml
[tool.poetry]
name = "django_icon"
version = "0.1.0"
description = "A simple icon field for Django models."
authors = ["Andre Guerra <andre.catarino.guerra@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.1"
Pillow = "^5.4"
python-intervals = "^1.8"

[tool.poetry.dev-dependencies]
coverage = "^4.5"
tox = "^3.7"
codecov = "^2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
